move `FeedImageMapper` into a separate file
